pyproject.toml
[project]
name = "tinyrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "jq>=1.11.0",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "numpy>=2.4.1",
    "openpyxl>=3.1.5",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.0",
    "unstructured>=0.18.27",
]
